# virtualkeypad/src/main/java/com/rfranco/virtualkeypad/controllers/UsersApiController.java
package com.rfranco.virtualkeypad.controllers;

import com.rfranco.virtualkeypad.autogenerated.api.UsersApi;
import com.rfranco.virtualkeypad.autogenerated.dtos.UserReponse;
import com.rfranco.virtualkeypad.services.UserService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.stereotype.Controller;
import org.springframework.web.bind.annotation.CrossOrigin;

import javax.validation.Valid;
import java.util.List;

@Controller
public class UsersApiController implements UsersApi{

    @Autowired
    private UserService userservice;

    @CrossOrigin
    @Override
    public ResponseEntity<Void> deleteUser(Integer userId) {
        this.userservice.deleteUser(userId);
        return new ResponseEntity(HttpStatus.OK);
    }

    @Override
    public ResponseEntity<UserReponse> getUser(Integer userId) {
        return new ResponseEntity(this.userservice.getUser(userId).get(), HttpStatus.OK);
    }

    @Override
    public ResponseEntity<List<UserReponse>> getUsers(@Valid String name, @Valid String lastname, @Valid String email) {
        return new ResponseEntity(this.userservice.getUsers(name, email, lastname),HttpStatus.OK);
    }

    @CrossOrigin
    @Override
    public ResponseEntity<Void> updateUser(Integer userId, @Valid UserReponse user) {
        this.userservice.updateUser(userId, user);
        return new ResponseEntity(HttpStatus.OK);
    }
}